FIxed sign-compare warning again

+++ cache/stack_simulator/dataset.hpp
@@ -284,7 +284,7 @@ bool ColMajorDataset<std::string>::import()
 
 	while (row_counter < max && std::getline(file, current_sample))
 	{
-		int pos_start = 0, pos_end = 0;
+		size_t pos_start = 0, pos_end = 0;
 		int column = 0, sample_column = 0;
 		
 		pos_end = current_sample.find(separator, pos_start);
@@ -342,7 +342,7 @@ bool ColMajorDataset<int>::import()
 
 	while (row_counter < max && std::getline(file, current_sample))
 	{
-		int pos_start = 0, pos_end = 0;
+		size_t pos_start = 0, pos_end = 0;
 		int column = 0, sample_column = 0;
 		
 		pos_end = current_sample.find(separator, pos_start);
